修改 install.sh

+++ resource/deploy/20211215/install.sh
@@ -123,10 +123,16 @@ if [ ! -d "${config_dir}" ]; then
 fi
 
 echo_info "------ 检查初始配置文件 ------ \n"
-cp "licence.key" "${config_dir}"
+if [ ! -f "${datarc_dir}/configs.py" ]; then
+  if [ -f "configs.py" ]; then
+    cp "licence.key" "${config_dir}"
+  fi
+fi
 
 if [ ! -f "${datarc_dir}/configs.py" ]; then
+  if [ -f "configs.py" ]; then
     cp "configs.py" "${config_dir}"
+  fi
 fi
 
 config_files=("update.sh" "docker-compose.yml" ".env" "configs.py" "licence.key")
